import and call function from another file in the same directory/package
