+++ .gitignore
@@ -1,3 +1,4 @@
 /target
 /firstapp.iml
 /guessing_game/guessing_game.iml
+/.idea/
