--- guessing_game/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "firstapp"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-rand = "0.8.5"
