--- guessing_game/src/main.rs
@@ -1,5 +1,0 @@
-use std::io;
-
-fn main() {
-
-}
